fix csg precision, improve texturing and entity inspector
- fixed: Merge and Bridge precision issues have been addressed, and both no longer fail on brushes with several slopey faces
- fixed: crash caused by the texture view when loading a second map
- texturing changes
	- support added for decimal texture shifts (didn't notice that GTKR supported both decimal shift and rotate when I added just rotate)
	- large offsets no longer clamped to sane ranges by repeatedly summing floats, for much less creep after huge translations
	- shift and rotate constrained in the surface editor to two decimal places (hundredths of a degree/pixel)
		- if you can look at a surface in-game and accurately tell if it's misaligned by thousandths of a pixel i will eat a shoe
	- fit now takes the sign of current scale values into account as well as rotation, so it no longer un-flips flipped textures
	- fractional fits will fit to the fractional range of the texture closest to existing alignment, so it no longer resets trim offsets
- raised command combine time from 400ms to 600ms
- modified entity inspector
	- spawnflags now go up to 16 since both Quoth and Copper use spawnflags above !Deathmatch for Coop flags
	- if an eclass doesn't specify more than 8 spawnflags, QE3 fills in the default !Skill flag names at parse time
	- added a third, even-wider layout that reserves most empty space for a full-height comment block

+++ CmdHollow.cpp
@@ -46,7 +46,7 @@ void CmdHollow::Do_Impl()
 		for (f = br->faces; f; f = f->fnext)
 		{
 			split = f->Clone();
-			move = f->plane.normal * (float)g_qeglobals.d_nGridSize;
+			move = f->plane.normal * (double)g_qeglobals.d_nGridSize;
 			split->plane.Translate(-move);
 
 			CSG::SplitBrushByFace(br, split, &front);
